Add files via upload
lstm_rnn_test is the main notebook where to train the model (approx 10 minutes to run) and visualise the results and data. For this you need the three packages called machine_learning_plotting/data/training to run the entire notebook. The three packages will be updated regulary and the main notebook will become clearer and more user-friendly soon :)